+++ lib/agent0/bin/fuzz_bots.py
@@ -8,13 +8,13 @@ import sys
 import rollbar
 from ethpy.hyperdrive.api import HyperdriveInterface
 from fixedpointmath import FixedPoint
+from hyperlogs.rollbar_utilities import initialize_rollbar
 from web3.types import RPCEndpoint
 
 from agent0 import initialize_accounts
 from agent0.base.config import AgentConfig, EnvironmentConfig
 from agent0.hyperdrive.exec import setup_and_run_agent_loop
 from agent0.hyperdrive.policies import Zoo
-from agent0.logging.rollbar_utilities import initialize_rollbar
 
 STOP_CHAIN_ON_CRASH = False
