+++ lib/hyperlogs/hyperlogs/rollbar_utilities.py
